# api/openapi.yaml
openapi: 3.0.3
info:
  title: supabackup API
  version: 0.1.0
  description: |
    supabackup control-plane API (Phase 1 scope: health, bootstrap, auth).

    Contract-first: this file is the single source of truth. Server types are
    generated with oapi-codegen (strict server), frontend types with
    openapi-typescript. CI fails if committed generated files drift from this
    contract. Generated types do NOT replace server-side runtime validation.
servers:
  - url: /api
    description: All routes are served under the /api base path by the server.
security:
  - cookieAuth: []
paths:
...
  /auth/bootstrap:
    post:
      operationId: postAuthBootstrap
      summary: Create the initial admin using a one-time token from the local CLI.
      description: |
        Fails with 403 if the token is unknown, used or expired; 409 if any
        admin user already exists. This endpoint can never be used by a
        public first visitor without the locally printed token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapRequest"
      responses:
        "201":
          description: Admin created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Invalid, used or expired bootstrap token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Already initialized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /auth/login:
    post:
      operationId: postAuthLogin
      summary: Log in with username and password.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in; session and CSRF cookies are set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Bad credentials (indistinguishable unknown-user vs wrong-password).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          description: Rate limited (too many attempts or consecutive failures).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/Internal"
        "503":
          $ref: "#/components/responses/Overloaded"
...
components:
  responses:
    PayloadTooLarge:
      description: Request body exceeds the allowed size.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: Content-Type must be application/json.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limited; retry later.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Internal:
      description: Unexpected internal error (nothing persisted, no state change reported as success).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Overloaded:
      description: Temporary overload (e.g. password-hashing budget exhausted); retry later.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb_session
      description: >-
        Server-side session. The cookie name is `sb_session` in plain-HTTP dev
        mode and `__Host-sb_session` when Secure cookies are enabled (production).
        State-changing requests additionally require the `X-CSRF-Token` header
        echoing the session-bound CSRF cookie.
  schemas:
...
    User:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        createdAt:
          type: integer
          format: int64
          description: Unix seconds.
    BootstrapRequest:
      type: object
      required: [token, username, password]
      properties:
        token:
          type: string
          description: One-time token printed by `supabackup bootstrap`.
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          format: password
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error class.
        message:
          type: string
          description: Human-readable, credential-free.